Update the cli package and the quickstart section of the docs
