+++ docs/src/SUMMARY.md
@@ -1,10 +1,12 @@
 # Summary
 
 - [Introduction](./introduction.md)
-  <!-- - [Principle](./principle.md) -->
-  <!-- - [Guide to fuzz my Tauri app](./guide_fuzz_my_app.md) -->
+  - [Prerequisites](./prerequisites.md)
+    <!-- - [Principle](./principle.md) -->
+    <!-- - [Guide to fuzz my Tauri app](./guide_fuzz_my_app.md) -->
 - [User Guide](./user_guide/intro.md)
-  - [Getting Started](./user_guide/getting_started.md)
+  - [Quick Start](./user_guide/quick_start.md)
+  - [Manual Steps](./user_guide/manual_fuzzing.md)
   - [Available Policies](./user_guide/available_policies.md)
   - [Write your Custom Policy](./user_guide/write_custom_policy.md)
   - [Fuzzer CLI](./user_guide/fuzzer_cli.md)
