+++ docs/src/prerequisites.md
@@ -1,0 +1,8 @@
+# Prerequisites
+
+> [!NOTE]
+> `AppFuzz Runtime` works with Tauri 2.x.
+
+- [Rust](https://www.rust-lang.org)
+- [LibAFL](https://github.com/AFLplusplus/LibAFL)
+- [Tauri Dependencies](https://tauri.app/start/prerequisites/), if you want to fuzz a Tauri application.
