+++ examples/mini-app/src-tauri/fuzz/.gitignore
@@ -1,0 +1,5 @@
+corpus_discovered
+*_solutions
+/target/
+
+
