+++ examples/mini-app/src-tauri/fuzz/build.rs
@@ -1,0 +1,8 @@
+// Copyright 2023-2024 CrabNebula Ltd., Alexandre Dang
+// SPDX-License-Identifier: PolyForm-Noncommercial-1.0.0
+
+/// Important: keep it here.
+/// Currently this is necessary for Windows
+fn main() {
+    tauri_build::build()
+}
