+++ examples/mini-app/src-tauri/fuzz/fuzzer_config.toml
@@ -1,0 +1,10 @@
+# Where to print the fuzzer outputs
+stdout = '/dev/stdout'
+# Number of cores used to fuzz
+nb_of_cores = 1
+# Directories containing starting input to start fuzzing
+corpus_input = []
+# Enable code coverage optimization
+with_coverage = false
+# Port used by the fuzzer broker
+broker_port = 8888
