+++ tests/fuzz-mini-app/Cargo.toml
@@ -7,19 +7,15 @@ edition = { workspace = true }
 license = { workspace = true }
 repository = { workspace = true }
 
-[build-dependencies]
-tauri-build = { workspace = true }
-
 [dependencies]
 mini-app = { workspace = true }
 policies = { workspace = true }
 appfuzz_rt = { workspace = true, features = ["tauri"] }
-frida-gum = { workspace = true }
+
 tauri = { workspace = true }
 tauri-utils = { workspace = true }
 libafl = { workspace = true }
 tauri-plugin-fs = { workspace = true}
 serde = { version = "1.0.197" }
 color-backtrace = "0.6.1"
-log = "0.4"
 env_logger = "0.11"
